Mix of language and IDE bug fixes
- Parselets fix for generating a change to 'b' in 'a.b'.  It was previously addi
ng an extra 'null' to the array which then would mess up style checking.  This s
howed up in the IDE rename feature when we went to re-annotate the model... the
styles would come back corrupt.

- Parselets fix for computing styles of 'string' valued elements under an Indexe
dChoice.   For 'spacing' we were previously returning the style for 'comment' ev
en for whitepsace which is wrong and inefficient.  Now we match the 'index' key
against the string and return the accurate 'child parselet' for a given string v
alue to fix this.

- Fixed problems with inactive types refering to active types - added diagnostic
s to catch this problem sooner.

- Fixed bug with the non-parameterized type version of 'getTypeForIdentifier' fo
r Selector expressions

- Fixed bug with type parameters with upper bounds

- Added code for "UpperBoundsTypeDeclaration" but then decided not to use it any
way

- Fixed bug with resolving type parameters mixing Number and a subclass of Numbe
r

- Reported error for @GetSet on non-field

- Fixed bugs with loading, refreshing type index.  To preserve the order of laye
rs and lazily load inactive layers, we need to keep a record of the layer order
in an index file.

- Began work on 'generateAllFiles' option - to improve the incremental compile.
 This is just a start.