Various bug fixes - scss in IDE, CFclass fixes
- added error for when a method that returns a value, has no return statement.  Cleaned up the
other return errors a little.

- fixed small bug with SC override grammar in partialValues parse-mode

- fixed CFClass bug - access restrictions were not being enforced for field lookups in CFclass

- sync manager - implement clone for cloneable types in a way that's portable to JS

- fixes for CSSLanguage

- CFClass - avoid resolving extend/implements types as source files if there's a .class file in the
classpath.  We don't yet support indexing and other features to allow CFClass files to extend files which
only exist in source.  It's a weird situation that should never happen under proper build rules anyway but
down the road, maybe we want to support this for the best tooling experience

- fixed bug in the convertToSrc process which happens when you are loading runtime types that have not been
compiled yet in that runtime.  Need to make sure we start the extended types and initialize the implements array.

- styling rules for selector expressions