reparse switch statements and resolve in ITypeDeclaration
There was an ArrayTypeDeclaration that hid a 'replacedType'

Reparse problems with the semantic value produced by the body for a switch statement.
It's a sequence that is repeating and which takes multiple child arrays, needing to
concatenate them to produce the final value.  We were not honoring the current childIndex
in the "removeFromSemanticValue" method and were not setting the "appendArray" flag in this case.