generate of empty finally and other fixes
error for duplicate methods

fixed bug with modify-extends where you add a
constructor that needs to call the super of the
new extends class.

try to avoid possible infinite loop in mis-reparsed ClassType

when generating a finallyStatement with no body, emptyValue in
the 'block' parselet was returning null - a BlockStatement with
a null statements.  So now the class BlockStatement implies a close
brace and that implies some content being generated.   We might need
a new property or something to differentiate this case from the
IfStatement where we use an empty 'else' keyword followed by a null
statement to be an empty value for that child parselet.  For now, it's
just relying on the fact that the skipped string value is the last
one in the parselet.  It's impossible then to have a null body so I think
this is a good default at least for the new attribute - ignoreEmptyProperties

more cases where we need to use ModelUtil.isCompiledClass - to fix
StrataCode features with cross compile option.

IDE exception fixes - null pointer, concurrent modification of peerSystems