Rework build process for better incremental builds
- There's a new preInit phase which detects changed files and stops any models from the
previous build which need to be restarted.  Then during init, all of the models are defined and
can be initialized.  Then all of the models are started + validated, then transformed.  We don't
do validate yet as a completely separate pass but that would not be hard to add now.

Before we relied on the 'replaced type' feature to retain the integrity of models that were not
reloaded.  We never explicitly stopped the old models and so it was harder to make things fully
reproducible so hopefully this will work better overall.