Refresh system bugs
New methods for allowing components to intercept 'type changed' events.
Using this for PageDispatcher so it can learn of page types being removed
or changed.   The addPage call might also get called when we update the type
but it's more reliable to use the TypeChangeListener

ResultSuffix annotation so we can learn a template's extension from the compiled
class.

Fixed bug with reparse - a NewExpression was inherting the partial value of
an IdentifierExpression which caused a corrupted data structure (so no longer
inheriting that method)

Refresh bugs:
  - clear the "lastStartedLayer" after we complete a build so we reliably
refresh types on a rebuild.  We don't want initializedInLayer to be set
for these types
  - passing a null "sinceLayer" will force a refresh
  - do not start types in updateType/updateModel.  We may have more than one type
we need to update so it's best to wait till all types are updates so when we
start references we get will be to new types.
  - run the updateInstance process after the build has completed.  Technically this
turned out not to be necessary but it seems like a better strategy than updating
the instances before the build.  We might be able to reload classes at some point
or need the transformed model in the update process.

Avoid double start method of Template for JS types

Fixed locking problem with saveTypeIndexChanges

Fixed "stale compiled model" warnings