fixed bug with misparsing multiple template statements as a tree
where <% a %> <% b %> was missing parsing as a template statement wrapping a glue expression,
not two template statements

- created a version of blockStatement/statement that does not have the glue and use that
for template statement bodies.  exprStatement still has glueExpression which in partial values
mode can parse as an expression without a semicolon.  It's awkward to clone
"expression" given how deep the parselet-tree is so for now we just disabled the partial values
parsing for exprStatement in the template language.  Ideally we'd have a way to disable a parselet
based on the tree.  We could use the SemanticContext for that but it's a heavy-weight solution to
something we could maybe fix more generally?

- support for cloning index choice parselets

- don't refresh non-started types

- new wrapper interfaces to support code formatting extensions in intelliJ