Added JSON/SCN switch - improved JS interface
- addition of JSON format - old way is now called "scn" with the -scn option to sc

- required adding more features to the JS api so we can run the deserialize operation on
the client (whereas before, we solved problems with intelligent code gen on the server)

- support hook for generic language conversion in format (to support sc -> js conversion)

- fixes to sync and html logging formats - verbose, trace, and traceAll

- fixes for serializing layers

- Fixed null pointer exception in reactivate for data binding

- added parameterTypes JSMethodSettings annotation for the rare case where we need to
change the type of the parameter for JS conversion (e.g. String.indexOf(int) is really indexOf(char) and it makes
a difference for the JS conversion)

- new dyn utilities for creating components, arrays with dynamic component type, 'applySyncLayer' so we can
abstract away the native SC language parsing from the sync system, getPropertyType - a hook so we can know
the type of a property in javascript for smarter deserialization.  Used only in a specific case right now but could be expanded
for more use cases down the road.

- js bugs:
   - fix for enum constant shadowed by method
   - make javascript variable '_c' shortcut for member definitions optional since it makes code harder to debug since you lose the type name
   - bug with missing clInit call for multi-leveled identifier expressions