three significant bug fixes
- bug fix for reparse - look for 'incomplete' parse-nodes that precede the changed region and
include them in the reparsed section.  Otherwise, we don't complete 'class foo {}' when adding 'extends bar' if at any point it turns into:
typeDeclaration[class foo, error:"extends", error:"{", error:"}"]

- support for intelliJ formatter for new expressions that are part of variable statements or fields

- change for 'stop' - if we are not started but are initialized, still go through teh stop procedure, just skip certain methods.  This caused problems with the reparse
tests and was not a problem per-se

- eliminate the removeTypesFromJSType - this causes problems with the new stop code - in general, trying to get annotations on types that
are being stopped is not working.   We probably should be doing the JS stop processing before we stop the type at all on all of the types in the type
hierarchy