Sync changes for keeping new and properties together
Needed a new wrapper for java.lang.Class objects going over the wire
and it seemed like an object with one property would be the best.

But properties get set after the instance has been restored.

Made some changes anyway to move primitive properties till right after the
'new' so objects are more in tact.  This does not work for cyclic references...
for now any references (not just cycles) are put at the end when we encounter
an on-demand object during serialization.  This could be improved by keeping track
of the pending objects and only serializing references to them after the object.

We also could just support deserializing cycles...

fixes to annotations

serialization fix for scn format - with new flag to control how top-level names
are used or not for json versus scn

fix to equalObjects - must use typeof as well as == for accurate JS comparison

fix for sync'd objects with no synchronized properties

fix for InstanceWrapper - it was sync'ing properties twice