disable reparseExtendedErrors and mark reparsed nodes as new
- during reparseExtendedErrors we were culling an ErrorParseNode that
had been reparsed and so was using a new index (re77)

During regressions, found that reparseExtendedErrors had another problem
with the fact that currentErrors is not being reset when we move the
current parse-index back and so during the reparse we did not get an
accurate partial value - this was inflating the # of parse-nodes parsed.
Disabled reparseExtendedErrors entirely - since we extend errors during
the original reparse.  Everything seems better and faster.