fixed several bugs, mostly dyn type related
- to get the new editor working in dynamic mode

- dyn stub constructor for components used to init the component
which wicket's PageCreator relied on via IComponent.class test which
incorrectly was just using 'new X' on component dyn stubs.  Now the
dyn stub implements IDynComponent - no methods since init etc are dyn
methods and we don't want all that extra code in the stub.

- fixes for how dyn constructors work when they are chained together.
There's an original type now in the ExecutionContext so we can always
construct the instance with the right type.  We also call constructDynInstance
in the constructor of the stub explicitly - rather than relying on
finding it during the super(x) chain process.