- various bug fixes found while automating runAllManual
- added dialog manager for sending answers to dialogs in the auto test

- do not use JLine's ! handling - setExpandEvents(false) as it messes up if (a != b)
on the command line

- eliminate some stale add messages

- lots of isDynamicType -> isDynamicNew since that call is used for both
dynamic types and dynamic new cases.

- do not exec stuff twice when the type is being updated - only exec the static members
during the update if the class has already been initialized

- do not re-resolve the extends type - this caused a bug after transform when the type had been
merged and one of the imports we needed to do the resolve was no longer available from the down
stream layer the type was merged into

- fixed bug with path types not being inherited properly... the original type was using
the absolute path which does not match the path type when it's added in the downstream layer.
so now we convert from an abs type to a relative one when we go through the parent layer.

- fixed implementation of live dynamic types - it should not be inherited through the parent
type, at least before we have tried to go through the base-type.