changes for syncing command line scripts - cmd.edit = true/false
- new script mode for command line - aka cmd.edit = false.   Currently all this effects is
property assignment and 'add block statement' to a type operation from the command line.  For those
operations, we can either modify the current type (edit = true) or just run the property assignment
without actually updating the body of the type declaration.   It's an important distinction that needs
some work I think... currently other edit operations are still allowed when edit = false... maybe
the option should be script = true so we don't contradict?   Or should we throw errors when you try to
edit in script-mode?  Or allow edits in some other form - like the temporary layer?

- execRuntine - new semantic node method to determine whether a given expression should be run
against a given runtime from the command line

- code update listeners - for real-time code-updates to the client

- js runtime - eval expression: ability to evaluate any expression from the command line