cmdline top-level member declaration support
Now the 'cmd' object starts out as a global, dynamic type which you can
use to add fields or methods to use as part of the script without having
to save them as normal types in the system.  This makes it a lot more convenient
to just prototype code or write code in a script.  No concerns about whether
the current layer is dynamic or compiled and you don't have to create a new type
just for your script code.

CmdClassDeclaration - the new TypeDeclaration for this 'cmd' class

cmd is now an IDynObject - so we can use the normal code for adding or replacing
fields, methods etc.

Needed to override the sync type for 'cmd' so it's not CmdClassDeclaration but
the regular old JLineInterpreter class

Fixed bug with resolveName and scope types

headless option improvement