fixed bug with IChildInit error
- Because we set jsLibFiles to map to a native JS class for Element, HtmlElement, HtmlPage, etc. we do not
process dependencies in those files during addTypeLib phase, but accidentally processed the IChildInit interface
while getting annotations later.  so this fix is two fold- one to add a way to manually specify real source
dependencies with the JSSettings(dependentTypes) annotation and also to avoid the accidental reference to
a file we encounter and include (but with an error due to "no type info") but were not explicitly told to include in the JS file.