Changes for binding events and synchronization
- we want to queue property change events during the 'applySync'
operation by default, but need to disable that when there's a remote method
call in the sync - i.e. sync before the method, run it in immediate mode, and/or
sync again afterwards.  Otherwise, things do not play out in the same
order.

Case in point - when we run the "click()" method on the client, we queue
the method return before we queue any side-effects of running that method
(the call to echoHello) - so we end up on the server possibly running the
echoHello after we get the return from the client.

- Also added a postListener to the RemoteResponse, which is run for the
remote method after the sync is complete.  We ended up not using that but
it seems like it could be useful.

- pass around BindingContext for the applySync operation