save/restore layer with cmd.includeSuper
importPackages - by default, import all layer packages so scripts are
less sensitive about the currentLayer.

switch current layer when we find a type in script mode, or in edit mode
when the package is incompatible with the current layer.

add/remove layer directory - (need to finish testing the remove
layer directory cases but at least the code is there now to handle
that case)

added 'singleton' support - before we use the dynamic type system,
look for an instance in the current scope context.  Not sure if this
is working or not, but eventually we need this for supporting
scripts that implement multi-application sharing of data