various changes addressing test script issues and bugs found
- maintain order of incoming events when queuing is enabled
- make sure current scope env is set when dispatching events
- for scopes which are not delivering events, update previous value right away.
since we've delivered the events to all of the listening scopes, we should immed
iately update the initial and previous value so subsequent changes are similarly propagated.
- added some messages for cases where we are applying changes from an inherited
scope.  not sure we do that anymore and we could speed up the code.
- added -ti option for running a test script and the command interpreter
- use "assignableTypes" not same types for setting current object
- fixed bug with going down a type hierarchy and then back up and not
ending up where you started because you navigated to a subclass.