basic query params, alt tags
query params now defined on HtmlPage on the server and Page on the
client - added comments about how to resolve this oddity.

PageInfo - this is only on the client now but mirrors the functionality
in PageDispatcher for managing the URL to page-type mapping.  Ideally we
could reconcile it so it's in a shared package with the server.  Awkwardly
we are implementing all of these types explicitly in tags.js because it's
more consistent with the packaging but that does not feel like a good excuse.

removed optimization for getElementById - we don't set children in the tag
for runtime Elements