fix for cloning Template's that have already been inited
rather than call convertToObject during the deepCopy method for the
Template, added a method to clone the model and reassign the types.
Not sure this is the best approach but seems a little faster and
more flexible?  It only comes up when we are cloning models in the
IDE for annotation purposes and the cached model has already been
registered with a PSI.  Not sure why that happens... maybe our cached
PSI has been invalidated?

added "not found error" flag for better IDE annotations