parsing changes for 'incompleteStatement'
- new parselet flag for partial-values-only - do not match a parselet unless
in the mode we use to parse for better errors and in the IDE

- this is the problem where you are typing in a class body and have a valid identifier
but which is not a complete statement.  we'd like to match a ClassType or something so
we have the right structure to do auto-complete in the IDE.  So we first tried parsing
the ClassType in the skipOnErrorParselet but although it worked (with some small changes
to how we create the value in OrderedChoice.parseExtendedErrors/reparseRepeatingChoice) it did
not set the parent node, start it or make it discoverable in the class body.  So now we
parse IncompleteStatement there which currently is just a wrapper around ClassType, but we can
use that to parse other constructs we need for better editing.

- removeParselet improvements

- bug with range calculation for utility used in formatting

- new ModelUtil.getParameterString for use in the IDE/doc as a way to identify methods clearly