fixed regression with layerResolve flag and buildSeparate layers
while we don't want to resolve source references in layers in general,
we still need to resolve them for buildSeparate layers.  We start those
layers along with the layers that depend on their compiled results before
we've compiled the buildSeparate layers.  There's other logic in there
to be sure we initialize these buildSeparate layers so we can find their
source while starting dependent layers.