schtml page element caching, server tag supports TodoList sample now, fixes for sync
cache attribute - enabled/disabled or unset - inherited in the tag hierarchy

attribute noCache -> stateless

property for each runtime - supportsTagCaching

wrapper classes that are server tags need to have a defined id so we can
set the innerHTML property to update the repeat body contents

serverTag fix for nested synchronized tags - when generating the body of an
enclosing tag, we don't need to set child change events but do need to update
the 'previous value' aka remote value for these properties like startTagTxt and
body which have changed indirectly via the parent.  otherwise, when we change
it again on the server, we might falsely think we don't have to send the change
to the client.

sync fixes:
   - reverse order for searching for SyncContext for a given instance to return
   the most specific SyncContext first.  Not sure why we were doing it the other way
   but need that for the new updateRemoteValue method on SyncManager which uses
   addPreviousValue to allow us to record a change made on the remote end that was "implicit" - i.e.
   a side-effect of another property change like the situation created when you have a hierarchical
   property that include a child property.