new way of implementing 'exec' for tag objects
permits 'overriding' of exec in a modify by using @sc.obj.Exec annotation

fixed bugs with Exec annotation

add -ve option - verboseExec

features derived from 'exec' must be applied during transform, not during model init
the 'init' phase must define a normal extends hierarchy and is the first time you have
access to the base types - modified or extends.   So we can't use annotations to exclude
generation of code, because those can be overridden in base-layers, not yet know at the
time of the initModel where we need to have types defined.

in the new model, this transformation happens during 'transform' - we can already exclude
types, mark expressions and statements as not available with "execForRuntime" so now
using this.

added hooks in Element for "addMixinProperties" and getExcludedStub so that we can control
the code-generation process during the transform phase based on attributes in the schhtml file.
mixin properties could have been done with a mix but was easier.  excluded stub is a candidate
for a future object template if anyone else needs to change that.