more serverTag bugs
- incremental invalidate of repeat tags for server (adapting the code
that's on the client already).  When the 'repeat' property changes, if
we can 'patch' the tag objects, we do not have to invalidate the parent.
For serverTags, this means delaying the change event for innerHTML of the
repeatWrapper until basically right before we check the bodyTxtValid flag
in fireChangedTagEvents.

- fixed bug with JS code-gen for server tags - don't treat excluded types
as an entry point.  Ideally we'd avoid adding them altogether but because
excluded gets set during the 'start' phase, a type is started before its
excluded.

- don't generate server tag stub for top-level type.  we weren't using it anyway
and that also messed up the code-generation for an 'object' serverTag on
the client.