fixed bugs with ExcludedStub
- since we are no longer generating the stub for the template if the entire thing is excluded,
when an inner class extends a top-level excluded type (e.g. HtmlPage.head extends Head) the Head
class was excluded and not found.   So in this case, we'll have the stub class extend the system
tag class which we know is defined.  We need the right base class so the id is formed properly for finding the
DOM element on the client.