fixes for server tags
- rather than explicitly firing change tag events in the page
rendering, do it in a 'doLater' so we can schedule these changes
for another sync even when changing something from the command line.

- fix for setVisible and validity.  since we don't validate invisible
children, we when going from invis to vis, we need to invalidate the
parent so the child gets re-rendered.

- don't use javascript js files unless there's an active js runtime
otherwise, we use the server tags files (stags.js)