java type system bugs
- do not tree abstract methods like interfaces for choosing the
most specific version of a method.  i.e. a subclass can have an abstract
clone method with a more specific return type.  We need to use the
abstract method, not the concrete one in that case.

- fix for Type<?> where Type is defined with Type<? extends Foo>.  We
were returning 'Object' as the type declaration but needed to return Foo.

- fix for resolveBaseTypeParameter for the case of finding the component type in:
Iterable<componentType> with a subclass in killbill called "javax.cache.Cache" which is defined like

    public interface Cache<K, V> extends Iterable<Cache.Entry<K, V>>...

Here were just needed to notice that the first extends type assigned a concrete type for the type
parameter we are looking for and just return that.