DynChild changes
when removing a child from a scope that has an outer object
should also be removed from the parent so the parent can invalidate
itself - e.g. when you have a  session scoped page that holds a reference
to a request scoped component, that request scope component gets removed
after the request completes, and recreated on the next request.  We need to
invalidate the body of the parent tag in that case.