syncTypeFilter, accessHook, sync-propagate bug and listenerAdded events
- added support for the syncTypeFilter - currently enabled by default.  For multi-page
web applications which might have some global or session scoped properties which are
shared, or not across the apps, need a way to direct changes only to the pages which have
that code.   Each page now compiles a "syncTypeFilter" - the list of all of the synchronized
types used by that page.  Only pages using the types receive the changes.   This can be turned
off globally right now but want to see if we can leave it on by default.

- added the "listenerAdded" event - which is used on the client side of the data binding library
but then ended up not using that approach.  Originally to fix the sync-propagate bug but that
was fixed instead by the accessHook so this listenerAdded event is in the code but not yet tested.

- the sync-propagate bug - needed to capture when a shared scope was used for the first time in a child
scope - i.e. a new window opens.  This requires that we refresh the page and we added a new "accessSyncInst"
hook for these shared scopes.  In general, the accessHook is like the customGetter/Setter hooks we already
use but is used in the 'else' case when the object already exists.