initTypeDecl method to fix weird regression bug
something changed in the scope templates which caused the class
to be initialized before it's body was defined which messed up
the computation of the 'dynamicType' flag.   In resolving an
attribute on a tag, we were calling definesMember which needed
the declaration type which was doing too much work by calling "init"
when all it really needed was to init the declaration type so that
code was broken out into a separate method.