don't use cache when timestamps are ==
low timestamp granularity on HFS for Mac.  So when refreshMQ or other
tests parse a file, then copy another one in the same second in place,
we might try to use the cached model when it's really invalid.  Ideally
we'd check the length and/or hash to be sure it's valid

started to added lazyParseNode feature