added disableProcessing and include/excludeRuntimes for LayerFile types
- for now, this is so that scsh files are only processed in the
server runtime, since that's where they will be executed. Although
I added the more general exclude/includeRuntimes, I ended up just
turning off the processing the JS runtime when !serverEnabled.
This logic was simpler and the procs are registered separately
for each layered system so it doesn't work to exclude the JS runtime
from the server layer - since the JS runtime won't have an instance
of that layer and so it won't get excluded there.

- cleaned up URLPath interface so that for URL params we can generate
the full Pattern object and skip optional params.