several bug fixes for dyn mode with doc sample
- added boolean getField parameter to getProperty methods

This goes along with setField which is in setProperty and helps us
implement overriding of inner child objects when overriding a compiled
property with a new dynamic type.  In getObjChildren
for a dyn type, we first try to get the field. If it's not null
check if it's a dynamic inner type and if so, use that mechanism
to create the child, then use the compiled setX method to set the field.

- added TypeSettings.dynAbstract annotation, useful for elements
with abstract='true' which are not actually made abstract, because
we want to use the base class for dynamic mode, to avoid extra stub
classes. Still need a way to identify these 'abstract' classes at
runtime or else we process annotations on them which we should not
like creating a JS module for HtmlPage once we get the compiled
class for it

- setNeedsDynamicStub change - performance problem?
  - we used to stop the propagation of the 'setNeedsDynamicStub' call
from the child to the parent when the parent already extended a dynamic
stub. But there's a weird case here - where we need the parent stub class
for defining an innerChild class... for example, if SomePage.schtml extends
HtmlPage which is compiled and overrides the 'head' tag to set the title
attribute, HtmlPage already is a dynamic stub but doesn't define it the head
innerclass as a dynamic stub.  Now we just create the extra stubs.