fixed bug with merging tags that do not use objects
- this required another change to getExtendsTypeDeclaration for Element
to resolve the extendsType - it is now getting bound a little too early
due to a check for exec flags - (CodeView.codeEditors extends CodeView.CodeEditor which is defined in editor.coreui and modified in editor.js.core)

- add includes for runtime/process constraints to layer detail string

- fixed another bug with including both runtime and process constraints

- added check for CancelledException when initializing a layer.
previously we'd create a dummy layer stub in this case which would
mess up the IDE when it happened.  Now, hopefully we just abort
the initialization of the layer which can be retried later.

- fixed -t option for runTests?

- readme cleanup