workaround for bug with replaced repository packages
right now we share RepositorySystem among LayeredSystems - and when
one RepositoryPackage gets added in a second system of the same name,
it uses the first system. This happens during 'init'. The second one
just indicates it's been replaced and points to the first one.

But the first one could later get excluded and removed from the system
which then unsets the layeredSystem pointer. We log an error now in
that case and check if we are being replaced by an excluded layer -
and if so, don't use it.

Maybe we should have separate RepositorySystem's for each LayeredSystem.
Then have a RepositoryCache which is shared.

We should probaby also remove the repository package when the layer is
stopped and removed.