several bug fixes found when testing the server side editor
- sync bug where class names matched but packages did not

- dynamic style sheets for server tags - add styleSheetPaths property
that collects the referenced .css files under the head tag.  It allows
the addServerTags method to perform sync on any sccss files referenced
in the page.

- added IPageDispatcher and IPageEntry for server tags style sheets -
but as a general a way to navigate from URLs to page types from core
functions

- parentNode parameter in constructor for creating repeat elements and
fix for ensuring unique ids in repeat tags

- added ModelUtil.getArrayNumDims - for array types, getting the number
of dimensions

- fixed bugs for JS arrays that are initialized.  We need to call
sc_newArray for arrays that are initialized and added a new JS core
function sc_initArray() to support that case.  added tests to test.js.arrayDim

- fixed bug with exceptions that occur inside of the loadLayerObject
method that were being hidden