SyncOptions to SyncPropOptions - server tag events
- flags to SyncProperties construtor now are default
sync prop options and serve as the default values
for the normal sync prop options.  Fixes a few problems
and cleans up the API.

- for clientToServer sync mode, we now still add the sync property
but set the new SYNC_RECEIVE_ONLY flag. This way we still use the
same data structure used for validating input requests but don't
add a sync listener on the server side

- bug in activateLayers

- build layer cleanup for the case where the build layer in the
main runtime is not in the JS runtime. we were using separate
build dirs and so the web/js files were not making it into the
right web folder.  Now we initialize the main system first and
use it's buildDir for the buildLayer's buildDir