data binding bug fixes, property names, setIndex fixes
for data binding, when there's an "a.b" make sure we validate
the value of binding 'a' when we are evaluating 'b'.

fixed bugs with array element binding - needs to return 'changed' value

fixed bugs with setX(int ix, val) for arrays. This cleared up some
problems in ArrBind that had been injected some time ago probably.

added setting for adding property names in the JS runtime