bug fix for dynamic eval of this.ix in setIx method
- need to make sure we resolve the field in this case so that
we don't try to recursively call the setIx method.  Same problem
for the constructor properties.  So the rule we came up with here is
that if we are in a named property method, we first look up the field,
then lookup the setX method. There are cases where we need to use the
this.x prefix to refer to a method call when there is no field.