- several changes to fix bugs found in JS IDE
- restore flag in setSemanticValue so we can do logic
in the semantic node when the parse node is restored
that's specific to the restore operation. Used only for
attributes in schtml where we want to keep track of the
index of the expression parsed from the code but seems like
a useful hook in general (despite being a lot of code to change)
It's possible the right fix here was to just restore the parse
node before the initModel but for dynamic templates that would
end up being more work.

- support for collecting type names of src files matching a
pattern to speed up goto class in the IDE

- fixes for constructor properties. needed to move the
initPropagateConstructor method to a separate lazy-init on the
first declaresConstructor call. Otherwise, we were initializing
dependent templates in a recursive way... basically initTypeInfo
is for setting up the type inheritance graph and initializing
the constructor properties requires starting stuff too early,
before all of that was set up.

- output() method should mark that it produces HTML (or at least
it's output should not be escaped)

- fix for sync repeat bug that cascaded into other fixes since
we were then hitting more cases where we reused elements.

- ensureStarted fix for convertToSrcReference that showed up when
changing the editor dependencies

- fix for dynamic constructors that showed up due to I think
changes made to repeat element type parameters

- fix for removing some errors in excluded types that showed up

- fix for repository packages in layers that are excluded, removed, replaced
etc. Need to track the list of repostiroy packages and the defining layer
so that we can always replace the removed one if necessary

- fix for the order of type resolution - need to sort the base layers
in the order in whcih tey are in the layer stack, not the order
they show up in the baseLayers list.

- fix for serializing modifier flags in the layer model

- fix for ensuring that we open layers to find the right type in the
inactive model

- verbose lock improvements

- fix for IDE - don't look up active layers... they can end up missing types in certain code views.

- improved errors for restoring parse node failures