data binding pendingChild, reverse nestedBinding flag fix and misc
need to propagate the pendingChild flag up the binding hierarchy through
the IBinding.getPropertyValue call so that we don't eval reverse bindings
'down' the tree once a remote result comes in. Effectively, it turns on the
cache binding flag which is normally not enabled for arguments to reverse
binding expressions but is needed here because we might need to gather up
multiple remote results before we can eval the reverse binding.

for a property assignment with a reverse-only assignment binding
as it's initializer, we were clearing the nestedbinding flag improperly
when initializing the statement again after adding it to the block statement
during the property assignment to assignment expression conversion.

started to implement static sync properties but then realized that we should
not tie the lifecycle of the values to the class and instead use scopeable
top-level objects (moved static variable in Blog to BlogManager class). the static
variable thing looked totally possible but would not create manageable apps so
for now it will just log an error when you try to do it.

some debug code for trying to eliminate a recursive infinite loop resolving
a model during the editing process

bug fix for sync system for a case where we were not setting the package properly
after serializing some dependencies